Constants cleanup
